test(adapters): share typed download client double
