+++ tests/unit/cli/conftest.py
@@ -14,12 +14,7 @@ import notebooklm.cli.context as context_module
 import notebooklm.cli.helpers as helpers_module
 import notebooklm.cli.resolve as resolve_module
 import notebooklm.cli.services.session_context as session_context_module
-from notebooklm.downloads import DOWNLOAD_REGISTRY, resolve_download_format
 from notebooklm.types import (
-    Artifact,
-    ArtifactDownloadListing,
-    ArtifactDownloadRequest,
-    ArtifactDownloadSelection,
     MindMapResult,
     ResearchSource,
     ResearchStart,
@@ -27,6 +22,7 @@ from notebooklm.types import (
     ResearchTask,
     SourceGuide,
 )
+from tests._helpers.downloads import configure_prepared_artifact_downloads
 
 
 @pytest.fixture(autouse=True)
@@ -295,45 +291,6 @@ class MockNote:
     def __init__(self, id: str, title: str = "Mock Note"):
         self.id = id
         self.title = title
-
-
-def configure_prepared_artifact_downloads(mock_client: MagicMock) -> None:
-    """Install the typed download contract on a client double's artifacts namespace."""
-
-    async def prepare_downloads(request: ArtifactDownloadRequest) -> ArtifactDownloadListing:
-        """Project the overridden typed list into public prepared identities."""
-        representation, output = resolve_download_format(request.kind, request.output_format)
-        artifacts = await mock_client.artifacts.list(request.notebook_id)
-        selections = tuple(
-            ArtifactDownloadSelection(
-                notebook_id=request.notebook_id,
-                artifact_id=artifact.id,
-                kind=artifact.kind,
-                title=artifact.title,
-                created_at=artifact.created_at,
-                last_modified_at=artifact.last_modified_at,
-                representation=representation,
-                extension=output.extension,
-                mime_type=output.mime_type,
-            )
-            for artifact in artifacts
-            if isinstance(artifact, Artifact)
-            and artifact.kind is request.kind
-            and artifact.is_completed
-        )
-        return ArtifactDownloadListing(selections, is_complete=True)
-
-    async def download(selection: ArtifactDownloadSelection, output_path: str) -> str:
-        """Bridge typed execution to a test's established per-kind callback."""
-        entry = next(item for item in DOWNLOAD_REGISTRY if item.kind is selection.kind)
-        legacy = getattr(mock_client.artifacts, entry.download_attr)
-        kwargs: dict[str, str] = {"artifact_id": selection.artifact_id}
-        if entry.format_kwarg:
-            kwargs[entry.format_kwarg] = selection.representation
-        return await legacy(selection.notebook_id, output_path, **kwargs)
-
-    mock_client.artifacts.prepare_downloads = AsyncMock(side_effect=prepare_downloads)
-    mock_client.artifacts.download = AsyncMock(side_effect=download)
 
 
 def create_mock_client():
